+++ src/components/Profile/ProfilePosts.jsx
@@ -6,12 +6,11 @@ import PostGalleryItem from './PostGalleryItem';
 import BottomSheetModal from '../../layouts/Modal/BottomSheetModal';
 import BottomSheetContent from '../../layouts/Modal/BottomSheetContent';
 import ConfirmModal from '../../layouts/Modal/ConfirmModal';
-import { ReactComponent as PostGalleryIcon } from '../../assets/icons/icon-post-album.svg';
-import { ReactComponent as PostListIcon } from '../../assets/icons/icon-post-list.svg';
 import useModal from '../../hooks/useModal';
 import { deletePostAPI, reportPostAPI } from '../../api/apis/post';
 import { userpostAPI } from '../../api/apis/post';
 import PostNone from './PostNone';
+import PostAlignButtons from './PostAlignButtons';
 
 const PostsContainer = styled.section`
   display: flex;
@@ -28,22 +27,6 @@ const PostAlignWrapper = styled.div`
   width: 100%;
   border-top: 0.5px solid var(--border-color);
   border-bottom: 0.5px solid var(--border-color);
-`;
-
-const PostsAlignRow = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  max-width: 390px;
-  width: 100%;
-  height: 44px;
-  padding-right: 10px;
-`;
-
-const AlignButton = styled.button.attrs({ type: 'button' })`
-  background: transparent;
-  border: none;
-  height: 26px;
 `;
 
 const PostGalleryView = styled.ul`
@@ -127,50 +110,11 @@ export default function ProfilePosts({ accountname, setIsPostLoading }) {
         <PostsContainer>
           <h2 className="a11y-hidden">프로필 포스트</h2>
           <PostAlignWrapper>
-            <PostsAlignRow>
-              <AlignButton onClick={handleListAlign}>
-                {isListView ? (
-                  <PostListIcon
-                    fill="var(--main-color)"
-                    stroke="var(--main-color)"
-                    aria-hidden={true}
-                    role="img"
-                  >
-                    <desc id="desc">리스트 뷰 활성화</desc>
-                  </PostListIcon>
-                ) : (
-                  <PostListIcon
-                    fill="var(--light-gray)"
-                    stroke="var(--light-gray)"
-                    aria-hidden={true}
-                    role="img"
-                  >
-                    <desc id="desc">리스트 뷰 비활성화</desc>
-                  </PostListIcon>
-                )}
-              </AlignButton>
-              <AlignButton onClick={handleGalleryAlign}>
-                {isListView ? (
-                  <PostGalleryIcon
-                    fill="var(--light-gray)"
-                    stroke="var(--light-gray)"
-                    aria-hidden={true}
-                    role="img"
-                  >
-                    <desc id="desc">갤러리 뷰 비활성화</desc>
-                  </PostGalleryIcon>
-                ) : (
-                  <PostGalleryIcon
-                    fill="var(--main-color)"
-                    stroke="var(--main-color)"
-                    aria-hidden={true}
-                    role="img"
-                  >
-                    <desc id="desc">갤러리 뷰 활성화</desc>
-                  </PostGalleryIcon>
-                )}
-              </AlignButton>
-            </PostsAlignRow>
+            <PostAlignButtons
+              isListView={isListView}
+              handleListAlign={handleListAlign}
+              handleGalleryAlign={handleGalleryAlign}
+            />
           </PostAlignWrapper>
           {isListView ? (
             <PostListView>
